requirements.txt
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.8.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
db==0.1.1
db-sqlite3==0.0.1
docstring_parser==0.16
duckdb==1.2.0
Faker==36.1.1
fastapi==0.115.8
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
Markdown==3.7
MarkupSafe==3.0.2
numpy==2.2.3
packaging==24.2
pillow==11.1.0
pybase64==1.4.0
pydantic==2.10.6
pydantic_core==2.27.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
